autodiscovery update
now works to the point that the entity is added in the MQTT integration but the device does not show up. The code in its current format would not even add the entity because of the device info being sent messes things up. Without it the entity would be added though.